175 multiplayer complete refactoring to new hostclient logic

+++ Game/Scripts/Multiplayer/Snapshot.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 
 public class Snapshot {
-    public int Tick;
+    public ulong Tick;
     public List<EntitySnapshot> Entities = new List<EntitySnapshot>();
-    public Snapshot(int tick) { Tick = tick; }
+    public Snapshot(ulong tick) { Tick = tick; }
 }
